default/requirements.txt
pynetworktables
dt_apriltags
opencv-contrib-python
numpy